temporarily keeping characters and persona prompts when refreshing pipeline

+++ pipeline/start_pipeline_afresh.py
@@ -12,9 +12,10 @@ def clear_output_folders():
   results_path = output_path.joinpath("results")
   for path in [output_path, characters_path, persona_prompts_path, attack_prompts_path, transcripts_path, results_path]:
     if path.exists():
-      for file in path.iterdir():
-        if file.is_file():
-          file.unlink()
+      if path not in [characters_path, persona_prompts_path]: #not clearing characters and persona prompts
+        for file in path.iterdir():
+          if file.is_file():
+            file.unlink()
     else:
       path.mkdir(parents=True, exist_ok=True)
